EMS/requirements.txt
streamlit
google-api-python-client
google-auth
pandas
apscheduler
google-auth-oauthlib
plotly
